rename libfreedo files
While not an ideal solution there are build environments which will put object files in a single directory with a name based on the original file. This can lead to collisions and broken builds. Prefixing the files with 'freedo' greatly minimizes the likelihood of such collisions.
