+++ libfreedo/freedo_bitop.h
