+++ libfreedo/freedo_cdrom.h
