+++ libfreedo/freedo_core.h
