+++ libfreedo/freedo_diag_port.h
