+++ libfreedo/freedo_madam.h
