+++ libfreedo/freedo_vdlp.h
