+++ libfreedo/freedo_xbus_cdrom_plugin.h
